Save work

+++ Diet.Core/Dtos/MealDto.cs
@@ -8,7 +8,7 @@ namespace Diet.Core.Dtos
 
         public string Name { get; set; }
 
-        public double Calories { get; set; }
+        public int Calories { get; set; }
 
         public DateTime DateTimeCreated { get; set; }
     }
